% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ROC.R
\name{AUC}
\alias{AUC}
\title{A function to calculate AUC using non-cumulative response rates.}
\usage{
AUC(cpr, car, byDR = FALSE)
}
\arguments{
\item{cpr}{a vector of cp id rate.}

\item{car}{a vector of ca id rate.}

\item{byDR}{Whether to order the ids by DR. Defaults to FALSE.}
}
\value{
Calculated AUC.
}
\description{
A function to calculate AUC using non-cumulative response rates.
}
