% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_auc.R
\name{data_cum}
\alias{data_cum}
\title{A function to generate cumulative id rates for both cp and ca lineups}
\usage{
data_cum(data, byDR = FALSE)
}
\arguments{
\item{data}{A matrix with both cp and ca id rates.}

\item{byDR}{Whether to order ids by diagnosticity ratios. Defaults to FALSE.}
}
\value{
A data matrix with cumulative cp and ca id rates.
}
\description{
A function to generate cumulative id rates for both cp and ca lineups
}
