% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ReCalcID.R
\name{IDadj}
\alias{IDadj}
\title{A function to adjust the id rates for ca lineups using the 1/m method.}
\usage{
IDadj(id, conf = NULL, fid, sid, size = 6)
}
\arguments{
\item{id}{ID rate vector.}

\item{conf}{Confidence levels for the id rate vector. Default to be NULL.}

\item{fid}{Mapping confidence levels from filler id.}

\item{sid}{To-be-matched confidence levels for suspect id. Must have equal length as fid.}

\item{size}{Size of the lineup. Defaults to 6.}
}
\description{
A function to adjust the id rates for ca lineups using the 1/m method.
}
\examples{
ca_id <- c(rep(0,3), rep(c(0.1, 0.15, 0.25), 2))
names(ca_id) <- paste0(rep(c("IDS", "IDF", "REJ"), each = 3), c("high", "medium", "low"))

fid_conf <- paste0("IDF", c("high", "medium", "low"))
sid_conf <- paste0("IDS", c("high", "medium", "low"))

IDadj(ca_id, fid = fid_conf, sid = sid_conf)
}
