% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_SIMUDATA.R
\name{response}
\alias{response}
\title{A function to calculate responses from simulated memory distribution}
\usage{
response(memory, criterion, id_criterion = NULL, suspect = TRUE)
}
\arguments{
\item{memory}{A simulated data matrix with suspect (column 1) and filler memory.}

\item{criterion}{A vector of judgment criterion. Must have odd number of elements if id_criterion is not specified.}

\item{id_criterion}{A number to define the criterion for id or rejection. Use the middle element of criterion if not specified.}

\item{suspect}{Whether there is a designated suspect. Defaults to TRUE.}
}
\value{
A data matrix of id responses (IDS, IDF, or REJ) and confidence levels.
}
\description{
A function to calculate responses from simulated memory distribution
}
\examples{
n_sim <- 10000
y_memory <- replicate(n_sim, rnorm(1, mean = 2))
x_memory <- replicate(n_sim, rnorm(1, mean = 0))
filler <- t(replicate(n_sim, rnorm(5, mean = 0)))

cp_memory <- cbind(y_memory, filler)
ca_memory <- cbind(x_memory, filler)

response(cp_memory, 1:3)
}
