% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\name{response_calculate}
\alias{response_calculate}
\title{A function to calculate responses from simulated memory distribution}
\usage{
response_calculate(memory, criterion, id_criterion = NULL, suspect = TRUE)
}
\arguments{
\item{memory}{A simulated data matrix with suspect (column 1) and filler memory.}

\item{criterion}{A vector of judgment criterion. Must have odd number of elements if id_criterion is not specified.}

\item{id_criterion}{A number to define the criterion for id or rejection. Use the middle element of criterion if not specified.}

\item{suspect}{Whether there is a designated suspect. Defaults to TRUE.}
}
\value{
A data matrix of id responses (IDS, IDF, or REJ) and confidence levels.
}
\description{
A function to calculate responses from simulated memory distribution
}
