% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_auc.R
\name{roc_auc}
\alias{roc_auc}
\title{A function to calculate AUC using non-cumulative response rates.}
\usage{
roc_auc(cpr, car, group = NULL, byDR = FALSE)
}
\arguments{
\item{cpr}{A vector of cp id rates.}

\item{car}{A vector of ca id rates.}

\item{group}{A vector indicating group membership. If specified, will calculate AUC by group.}

\item{byDR}{Whether to order ids by diagnosticity ratios. Defaults to FALSE.}
}
\value{
Area under the curve.
}
\description{
A function to calculate AUC using non-cumulative response rates.
}
