% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ROC.R
\name{ROCarea}
\alias{ROCarea}
\title{A function to calculate AUC}
\usage{
ROCarea(cp_cum, ca_cum)
}
\arguments{
\item{cp_cum}{a vector of cumulative cp id rate.}

\item{ca_cum}{a vector of cumulative ca id rate.}
}
\value{
Calculated AUC.
}
\description{
A function to calculate AUC
}
