% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ROC.R
\name{ROCplot}
\alias{ROCplot}
\title{A function to plot the cumulative ROC curve.}
\usage{
ROCplot(
  cpr,
  car,
  cumulative = FALSE,
  byDR = FALSE,
  overlay = FALSE,
  cumuout = FALSE,
  ...
)
}
\arguments{
\item{cpr}{A vector of cp id rate}

\item{car}{A vector of ca id rate}

\item{cumulative}{Whether the id rates are cumulative. Defaults to FALSE.}

\item{byDR}{Whether to order the ids by DR. Defaults to FALSE.}

\item{overlay}{Whether to overlay the ROC curve on an existing plot. Defaults to FALSE.}

\item{cumuout}{Whether to output cumulative id rates. Defaults to be FALSE.}
}
\value{
An ROC plot and a data matrix of cumulative id rates used to create the plot.
}
\description{
A function to plot the cumulative ROC curve.
}
