% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot.R
\name{roc_plot}
\alias{roc_plot}
\title{A function to plot the cumulative ROC curve.}
\usage{
roc_plot(
  cpr,
  car,
  group = NULL,
  byDR = FALSE,
  xlab = "Innocent suspect rate",
  ylab = "Guilty suspect rate",
  ...
)
}
\arguments{
\item{cpr}{A vector of cp id rate}

\item{car}{A vector of ca id rate}

\item{group}{Grouping variable to indicate group membership. Will create an ROC curve and calculate AUC for each group.}

\item{byDR}{Whether to order the ids by DR. Defaults to FALSE.}
}
\value{
NULL. Generate an ROC plot and auc as side effects.
}
\description{
A function to plot the cumulative ROC curve.
}
